Merge Develop to Freeze

+++ plugins/org.eclipse.fordiac.ide.model.commands/src/org/eclipse/fordiac/ide/model/commands/change/ChangeCommentBoundsCommand.java
@@ -25,7 +25,7 @@ public class ChangeCommentBoundsCommand extends AbstractChangeContainerBoundsCom
 	}
 
 	@Override
-	protected void updateSize(final int width, final int height) {
+	protected void updateSize(final double width, final double height) {
 		getTarget().setWidth(width);
 		getTarget().setHeight(height);
 	}
