+++ plugins/org.eclipse.fordiac.ide.model/src-gen/org/eclipse/fordiac/ide/model/libraryElement/Comment.java
@@ -40,12 +40,12 @@ public interface Comment extends FBNetworkElement {
 	 * <!-- begin-user-doc -->
 	 * <!-- end-user-doc -->
 	 * @return the value of the '<em>Width</em>' attribute.
-	 * @see #setWidth(int)
+	 * @see #setWidth(double)
 	 * @see org.eclipse.fordiac.ide.model.libraryElement.LibraryElementPackage#getComment_Width()
-	 * @model default="200" dataType="org.eclipse.emf.ecore.xml.type.Int"
+	 * @model default="200" dataType="org.eclipse.emf.ecore.xml.type.Double"
 	 * @generated
 	 */
-	int getWidth();
+	double getWidth();
 
 	/**
 	 * Sets the value of the '{@link org.eclipse.fordiac.ide.model.libraryElement.Comment#getWidth <em>Width</em>}' attribute.
@@ -55,7 +55,7 @@ public interface Comment extends FBNetworkElement {
 	 * @see #getWidth()
 	 * @generated
 	 */
-	void setWidth(int value);
+	void setWidth(double value);
 
 	/**
 	 * Returns the value of the '<em><b>Height</b></em>' attribute.
@@ -63,12 +63,12 @@ public interface Comment extends FBNetworkElement {
 	 * <!-- begin-user-doc -->
 	 * <!-- end-user-doc -->
 	 * @return the value of the '<em>Height</em>' attribute.
-	 * @see #setHeight(int)
+	 * @see #setHeight(double)
 	 * @see org.eclipse.fordiac.ide.model.libraryElement.LibraryElementPackage#getComment_Height()
-	 * @model default="100" dataType="org.eclipse.emf.ecore.xml.type.Int"
+	 * @model default="100" dataType="org.eclipse.emf.ecore.xml.type.Double"
 	 * @generated
 	 */
-	int getHeight();
+	double getHeight();
 
 	/**
 	 * Sets the value of the '{@link org.eclipse.fordiac.ide.model.libraryElement.Comment#getHeight <em>Height</em>}' attribute.
@@ -78,7 +78,7 @@ public interface Comment extends FBNetworkElement {
 	 * @see #getHeight()
 	 * @generated
 	 */
-	void setHeight(int value);
+	void setHeight(double value);
 
 	/**
 	 * <!-- begin-user-doc -->
